requirements.txt
flask
requests
humanize
bbcode
Pillow
piexif
uwsgi